Fixed item limit bug and made read/write limits floating point

+++ src/main/scala/com/audienceproject/spark/dynamodb/connector/DynamoConnector.scala
@@ -45,7 +45,7 @@ private[dynamodb] trait DynamoConnector {
 
     val keySchema: KeySchema
 
-    val readLimit: Int
+    val readLimit: Double
 
     val itemLimit: Int
